requirements.txt
gitpython==3.1.42
pydantic==2.5.3
tree-sitter==0.20.4
tree-sitter-languages==1.10.2
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
